# app/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/login/access-token":{"post":{"tags":["login"],"summary":"Login Access Token","operationId":"login_access_token_api_login_access_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_api_login_access_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/employee/organization":{"get":{"tags":["users"],"summary":"Get Organization","operationId":"get_organization_api_employee_organization_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployerOrganization"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["users"],"summary":"Create Organization","operationId":"create_organization_api_employee_organization_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployerOrganizationCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployerOrganization"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/employee/{employee_id}":{"get":{"tags":["users"],"summary":"Get Employee","operationId":"get_employee_api_employee__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/employee/organization/{organization_id}":{"get":{"tags":["users"],"summary":"Get Employees By Organization","operationId":"get_employees_by_organization_api_employee_organization__organization_id__get","parameters":[{"name":"organization_id","in":"path","required":true,"schema":{"type":"string","title":"Organization Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationEmployees"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/employee/":{"post":{"tags":["users"],"summary":"Create Employee","operationId":"create_employee_api_employee__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Delete Employee","operationId":"delete_employee_api_employee__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"query","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tips/tips":{"get":{"tags":["tips"],"summary":"Get Total Tips","operationId":"get_total_tips_api_tips_tips_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/tips/add_tip":{"post":{"tags":["tips"],"summary":"Add Tip","operationId":"add_tip_api_tips_add_tip_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TipCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_access_token_api_login_access_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_access_token_api_login_access_token_post"},"Employee":{"properties":{"name":{"type":"string","maxLength":50,"title":"Name"},"is_active":{"type":"boolean","title":"Is Active","default":true},"upi_id":{"type":"string","maxLength":50,"title":"Upi Id"},"employee_id":{"type":"string","title":"Employee Id"},"org_id":{"type":"string","title":"Org Id"}},"type":"object","required":["name","upi_id","org_id"],"title":"Employee"},"EmployeeCreate":{"properties":{"name":{"type":"string","maxLength":50,"title":"Name"},"is_active":{"type":"boolean","title":"Is Active","default":true},"upi_id":{"type":"string","maxLength":50,"title":"Upi Id"},"organization_id":{"type":"string","title":"Organization Id"}},"type":"object","required":["name","upi_id","organization_id"],"title":"EmployeeCreate"},"EmployerOrganization":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"organization_name":{"type":"string","maxLength":50,"title":"Organization Name"},"organization_description":{"type":"string","maxLength":50,"title":"Organization Description"},"is_active":{"type":"boolean","title":"Is Active","default":true},"organization_id":{"type":"string","title":"Organization Id"},"hashed_password":{"type":"string","title":"Hashed Password"},"total_tips":{"type":"integer","title":"Total Tips","default":0}},"type":"object","required":["email","organization_name","organization_description","hashed_password"],"title":"EmployerOrganization"},"EmployerOrganizationCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"organization_name":{"type":"string","maxLength":50,"title":"Organization Name"},"organization_description":{"type":"string","maxLength":50,"title":"Organization Description"},"is_active":{"type":"boolean","title":"Is Active","default":true},"password":{"type":"string","maxLength":100,"title":"Password"}},"type":"object","required":["email","organization_name","organization_description","password"],"title":"EmployerOrganizationCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrganizationEmployees":{"properties":{"org_id":{"type":"string","title":"Org Id"},"organization_name":{"type":"string","title":"Organization Name"},"organization_description":{"type":"string","title":"Organization Description"},"data":{"items":{"$ref":"#/components/schemas/Employee"},"type":"array","title":"Data"}},"type":"object","required":["org_id","organization_name","organization_description","data"],"title":"OrganizationEmployees"},"TipCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"amount":{"type":"integer","title":"Amount"}},"type":"object","required":["employee_id","amount"],"title":"TipCreate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/login/access-token"}}}}}}